dashboard/requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.17.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
boto3==1.34.0 